SVG modification for icon in modules table and etc
