+++ apps/server/src/models/schema/modules.ts
@@ -7,6 +7,7 @@ import {
   pgTable
 } from 'drizzle-orm/pg-core';
 import { createInsertSchema, createSelectSchema } from 'drizzle-zod';
+import * as z from 'zod';
 
 export const ModuleSchema = pgTable('modules', {
   id: uuid('id').notNull().defaultRandom().primaryKey(),
@@ -23,6 +24,13 @@ export const ModuleSchema = pgTable('modules', {
 });
 
 export type ModuleInsert = typeof ModuleSchema.$inferInsert;
-export const ModuleInsertSchema = createInsertSchema(ModuleSchema);
+export const ModuleInsertSchema = createInsertSchema(ModuleSchema).merge(
+  z.object({
+    icon: z
+      .string()
+      .startsWith('<svg xmlns="http://www.w3.org/2000/svg"')
+      .endsWith('</svg>')
+  })
+);
 export type ModuleSelect = typeof ModuleSchema.$inferSelect;
 export const ModuleSelectSchema = createSelectSchema(ModuleSchema);
