+++ apps/server/src/services/UserService.ts
@@ -70,7 +70,8 @@ class UserService {
                 module: {
                   columns: {
                     name: true,
-                    id: true
+                    id: true,
+                    icon: true
                   }
                 }
               }
